use uint256 instead of uint (#21)

+++ contracts/Lottery.sol
@@ -4,42 +4,42 @@ import "./RandomNumberOracle.sol";
 
 contract Lottery {
     
-    uint constant GAME_LENGTH = 3;          // number of blocks
-    uint constant DRAW_PERIOD = 1;          // number of blocks
-    uint constant TICKET_PRICE = 1 ether;
-    uint constant REFUND_AMOUNT = 0.1 ether;
-    uint constant MIN_NUMBER = 1;
-    uint constant MAX_NUMBER = 5;
-    uint constant MAX_AMOUNT_TICKETS = 3;
-    uint constant MAX_PARTICIPANTS = 50;
+    uint256 constant GAME_LENGTH = 3;          // number of blocks
+    uint256 constant DRAW_PERIOD = 1;          // number of blocks
+    uint256 constant TICKET_PRICE = 1 ether;
+    uint256 constant REFUND_AMOUNT = 0.1 ether;
+    uint256 constant MIN_NUMBER = 1;
+    uint256 constant MAX_NUMBER = 5;
+    uint256 constant MAX_AMOUNT_TICKETS = 3;
+    uint256 constant MAX_PARTICIPANTS = 50;
     
     struct Game {
-        uint startBlock;
-        uint endBlock;
-        uint drawBlock;
-        uint luckyNumber;
-        mapping (uint => Participant) participants;
-        mapping (uint => address payable) winners;
+        uint256 startBlock;
+        uint256 endBlock;
+        uint256 drawBlock;
+        uint256 luckyNumber;
+        mapping (uint256 => Participant) participants;
+        mapping (uint256 => address payable) winners;
         
         // keep track of number of participants and winners to iterate
-        uint numberOfParticipants;
-        uint numberOfWinners;
+        uint256 numberOfParticipants;
+        uint256 numberOfWinners;
         
         // keep track of jackpot
-        uint jackpot;
+        uint256 jackpot;
     }
     
     struct Participant {
         address payable addr;
-        uint[] numbers;
+        uint256[] numbers;
     }
     
     Game public currentGame;
     
     RandomNumberOracle oracle;
     
-    mapping (uint => Game) public finishedGames;
-    uint numberOfGames = 0;
+    mapping (uint256 => Game) public finishedGames;
+    uint256 numberOfGames = 0;
 
     constructor() public {
         currentGame = createNewGame(GAME_LENGTH);
@@ -50,7 +50,7 @@ contract Lottery {
     // pubic functions
     // /////////////////////////////////////////////////////////////////
     
-    function buyTicket(uint number) public payable {
+    function buyTicket(uint256 number) public payable {
         // verify that enough ether was sent
         require(msg.value == TICKET_PRICE);
         
@@ -65,7 +65,7 @@ contract Lottery {
         
         // record the sender's address and the received number
         bool foundParticipant = false;
-        for (uint i=0; i<currentGame.numberOfParticipants; i++) {
+        for (uint256 i=0; i<currentGame.numberOfParticipants; i++) {
             // add number to existing participant
             if (currentGame.participants[i].addr == msg.sender) {
                 
@@ -83,7 +83,7 @@ contract Lottery {
             require(currentGame.numberOfParticipants < MAX_PARTICIPANTS);
             
             // create new participant
-            uint[] memory numbers = new uint[](1);
+            uint256[] memory numbers = new uint256[](1);
             numbers[0] = number;
             
             Participant memory p = Participant({
@@ -108,8 +108,8 @@ contract Lottery {
         // refund caller and payout winners
         if (currentGame.numberOfParticipants > 0) {
             // determine winners
-            for (uint i=0; i<currentGame.numberOfParticipants; i++) {
-                for (uint j=0; j<currentGame.participants[i].numbers.length; j++) {
+            for (uint256 i=0; i<currentGame.numberOfParticipants; i++) {
+                for (uint256 j=0; j<currentGame.participants[i].numbers.length; j++) {
                     if (currentGame.participants[i].numbers[j] == currentGame.luckyNumber) {
                         currentGame.winners[currentGame.numberOfWinners++] = currentGame.participants[i].addr;
                         break;
@@ -121,7 +121,7 @@ contract Lottery {
             msg.sender.transfer(REFUND_AMOUNT);
         
             // payout winners
-            for (uint i=0; i<currentGame.numberOfWinners; i++) {
+            for (uint256 i=0; i<currentGame.numberOfWinners; i++) {
                 currentGame.winners[i].transfer(currentGame.jackpot / currentGame.numberOfWinners);
             }
         }
@@ -134,10 +134,10 @@ contract Lottery {
     }
     
     // returns the tickets (= numbers) in the current game associated with the sender's address
-    function getMyTickets() public view returns(uint[] memory) {
-        uint[] memory numbers;
+    function getMyTickets() public view returns(uint256[] memory) {
+        uint256[] memory numbers;
             
-        for (uint i=0; i<currentGame.numberOfParticipants; i++) {
+        for (uint256 i=0; i<currentGame.numberOfParticipants; i++) {
             if (currentGame.participants[i].addr == msg.sender) {
                 numbers = currentGame.participants[i].numbers;
                 break;
@@ -147,7 +147,7 @@ contract Lottery {
     }
 
     // return the winnable amount in the ongoing game
-    function getJackpot() public view returns(uint) {
+    function getJackpot() public view returns(uint256) {
         if (address(this).balance >= REFUND_AMOUNT) {
             return address(this).balance - REFUND_AMOUNT;
         } else {
@@ -176,32 +176,32 @@ contract Lottery {
     }
     
     // returns minimum required number to buy a ticket (inclusive)
-    function getMinNumber() public pure returns(uint) {
+    function getMinNumber() public pure returns(uint256) {
         return MIN_NUMBER;
     }
     
     // returns maximum allowed number to buy a ticket (inclusive)
-    function getMaxNumber() public pure returns(uint) {
+    function getMaxNumber() public pure returns(uint256) {
         return MAX_NUMBER;
     }
     
-    function getNumberOfFinishedGames() public view returns(uint) {
+    function getNumberOfFinishedGames() public view returns(uint256) {
         return numberOfGames;
     }
     
-    function getStartBlockOfCurrentGame() public view returns(uint) {
+    function getStartBlockOfCurrentGame() public view returns(uint256) {
         return currentGame.startBlock;
     }
     
-    function getEndBlockOfCurrentGame() public view returns(uint) {
+    function getEndBlockOfCurrentGame() public view returns(uint256) {
         return currentGame.endBlock;
     }
     
-    function getDrawBlockOfCurrentGame() public view returns(uint){
+    function getDrawBlockOfCurrentGame() public view returns(uint256){
         return currentGame.drawBlock;
     }
     
-    function getCurrentBlock() public view returns(uint) {
+    function getCurrentBlock() public view returns(uint256) {
         return block.number;
     }
     
@@ -213,7 +213,7 @@ contract Lottery {
     // private functions
     // /////////////////////////////////////////////////////////////////
     
-    function createNewGame(uint gameLength) private view returns(Game memory) {
+    function createNewGame(uint256 gameLength) private view returns(Game memory) {
         Game memory newGame = Game({
             startBlock: block.number,
             endBlock: block.number + gameLength,
